requirements.txt
altgraph==0.17
ansi2html==1.5.2
ansicolors==1.1.8
appdirs==1.4.4
attrs==18.2.0
black==20.8b1
cefpython3==66.0
click==7.1.2
fdfgen==0.16.1
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
markdown2==2.3.9
MarkupSafe==1.1.1
mypy-extensions==0.4.3
npyscreen==4.10.5
pathspec==0.8.0
pdflatex==0.1.3
pdfrw==0.4
pefile==2019.4.18
pycparser==2.20
Pygments==2.7.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pypdftk==0.4
pythonnet==2.5.1
pywin32-ctypes==0.2.0
regex==2020.9.27
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicategories==0.1.0
Werkzeug==1.0.1
windows-curses==2.1.0
